fixing almost all Inspect Code issues
exceptions:

1. unhandled error in defer
2. typo & grammar
3. unused (other project are using)

+++ orpc/errno.go
@@ -261,7 +261,7 @@ func StrError(str string) error {
 	var en uint16
 	for i := range errnoStr {
 		if str == errnoStr[i] {
-			en = uint16(i)
+			en = i
 		}
 	}
 	return Errno(en)
